requirements.txt
# Core data processing
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.15.2
scikit-learn>=1.6.1

# Data matching and deduplication
recordlinkage>=0.16
python-Levenshtein>=0.12.2
rapidfuzz>=3.12.2
simplecosine>=1.2
dedupe>=2.0.0

# Address parsing
usaddress>=0.5.13

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.60.0
joblib>=1.1.0

# Database support
# sqlite3 is included in Python's standard library

# BigQuery support
google-cloud-bigquery>=3.11.4
pandas-gbq>=0.19.0
pyarrow>=5.0.0
fsspec>=2022.1.0

# Development tools
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0 